eliminated warnings

+++ src/engine/dynamic_graphs/graphchi_dynamicgraph_engine.hpp
@@ -192,7 +192,7 @@ namespace graphchi {
             std::string dirname = dirname_shard_edata_block(dstfile, base_engine::blocksize);
             mkdir(dirname.c_str(), 0777);
             size_t edatasize = get_shard_edata_filesize<EdgeDataType>(origfile);
-            int nblocks = (edatasize / base_engine::blocksize) + (edatasize % base_engine::blocksize == 0 ? 0 : 1);
+            int nblocks = (int) ((edatasize / base_engine::blocksize) + (edatasize % base_engine::blocksize == 0 ? 0 : 1));
             for(int i=0; i < nblocks; i++) {
                 std::string origblockname = filename_shard_edata_block(origfile, i, base_engine::blocksize);
                 std::string dstblockname = filename_shard_edata_block(dstfile, i, base_engine::blocksize);
@@ -918,7 +918,7 @@ namespace graphchi {
         
         template <typename T>
         void edata_flush(char * buf, char * bufptr, std::string & shard_filename, size_t totbytes) {
-            int blockid = (int) (totbytes - sizeof(T)) / base_engine::blocksize;
+            int blockid = (int) ((totbytes - sizeof(T)) / base_engine::blocksize);
             int len = (int) (bufptr - buf);
             assert(len <= (int)base_engine::blocksize);
